requirements.txt
-e .

pytest
pytest-cov
pytest-sugar
pytest_click
pylint
autopep8
coverage

